etc: separate types on file
