+++ src/components/Header/Mobile/MobileHeader.tsx
@@ -6,7 +6,7 @@ import menuBar from '@src/assets/icons/menuBar.svg';
 import xButton from '@src/assets/icons/xButton.png';
 import logoIcon from '@src/assets/sopt/logo.png';
 import { Condition } from '@src/lib';
-import { MenuList } from '../Header';
+import { MenuList } from '../types';
 import HeaderMenu from './HeaderMenu';
 
 export type MenuType = 'idle' | 'open' | 'close';
